[SERVER] iscsi: Refactor receive function and PDU handling
- Fold header/data handling into one function
  This uncovered a few redundant checks
  and makes it easier to reason about control flow

- Make all iscsi_pdu stack-allocated
  This greatly reduces the number of malloc and free calls
  during normal operation, lowers the risk of memory management
  bugs, and potentially increases performance in high concurrency
  scenarios.

+++ inc/dnbd3/shared/sockhelper.h
@@ -120,6 +120,6 @@ ssize_t sock_recv(const int sock, void *buffer, const size_t len);
 /**
  * Send a desired number of nullbytes to socket.
  */
-bool sock_sendPadding(int fd, uint32_t bytes);
+bool sock_sendPadding(int fd, size_t bytes);
 
 #endif /* SOCKHELPER_H_ */
